Fixes for MSVC (ReadWriteBuffer fails)

+++ tests/buffer/buffer_set_final_data_1.cpp
@@ -56,7 +56,7 @@ int test_main(int argc, char *argv[]) {
       queue {}.submit([&](handler &cgh) {
           auto write = buff.get_access<access::mode::write>(cgh);
 
-          cgh.parallel_for<class generate>(range<1> { N },
+          cgh.parallel_for<class generate>(range<1> { size_t(N) }, // MSVC: error C2398: Element '1': conversion from 'const int' to 'size_t' requires a narrowing conversion
                                            [=] (id<1> index) {
                                              write[index] = index;
                                            });
